frontend/requirements.txt
kombu==4.2.1
bcrypt==3.1.4
Flask==0.12.2
Flask-Babel==0.11.2
Flask-BabelEx==0.9.3
Flask-Bootstrap4==4.0.2
Flask-Caching==1.4.0
Flask-Compress==1.4.0
Flask-Hashing==1.1
Flask-JWT==0.3.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
flask-nav==0.6
Flask-Principal==0.4.0
Flask-Restless==0.17.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-Table==0.5.0
Flask-WTF==0.14.2
WTForms==2.1
SQLAlchemy==1.2.5
PyYAML==3.13
requests==2.19.1